Мой первый коммит.

+++ Converter.java
@@ -1,6 +1,5 @@
 public class Converter {
-
- void convert(int sum, int kcal, int km) {
+ void convert(int sum, double kcal, double km) {
  System.out.println("Сожённых килокалорий - " + sum*kcal);
     System.out.println("Пройдено километров - " + sum*km);
     }
